+++ src/Component/Analyser/ClassAnalyser.php
@@ -1,29 +1,21 @@
 <?php
+
 /**
  * Copyright (C) GrizzIT, Inc. All rights reserved.
  * See LICENSE for license details.
  */
+
 namespace GrizzIt\ObjectFactory\Component\Analyser;
 
 use ReflectionClass;
 use GrizzIt\Storage\Common\StorageInterface;
 use GrizzIt\ObjectFactory\Common\ClassAnalyserInterface;
+use GrizzIt\ObjectFactory\Common\MethodReflectorInterface;
+use GrizzIt\ObjectFactory\Component\Reflector\MethodReflector;
 use GrizzIt\ObjectFactory\Exception\NonInstantiableClassException;
 
 class ClassAnalyser implements ClassAnalyserInterface
 {
-    /**
-     * A translation array to equalize the expected types of the gettype method.
-     *
-     * @var array
-     */
-    private $translations = [
-        'bool' => 'boolean',
-        'int' => 'integer',
-        'float' => 'double',
-        'callable' => 'array',
-    ];
-
     /**
      * Contains the previously analysed classes.
      *
@@ -32,13 +24,26 @@ class ClassAnalyser implements ClassAnalyserInterface
     private $analysisStorage;
 
     /**
+     * Contains the method reflector.
+     *
+     * @var MethodReflectorInterface
+     */
+    private $methodReflector;
+
+    /**
      * Constructor
      *
      * @param StorageInterface $analysisStorage
+     * @param MethodReflectorInterface $methodReflector
      */
-    public function __construct(StorageInterface $analysisStorage)
-    {
+    public function __construct(
+        StorageInterface $analysisStorage,
+        MethodReflectorInterface $methodReflector = null
+    ) {
         $this->analysisStorage = $analysisStorage;
+        $this->methodReflector = $methodReflector ?? new MethodReflector(
+            $analysisStorage
+        );
     }
 
     /**
@@ -51,57 +56,20 @@ class ClassAnalyser implements ClassAnalyserInterface
     public function analyse(string $class): array
     {
         if (!$this->analysisStorage->has($class)) {
+            $reflection = new ReflectionClass($class);
+
+            if (!$reflection->isInstantiable()) {
+                throw new NonInstantiableClassException(
+                    $class
+                );
+            }
+
             $this->analysisStorage->set(
                 $class,
-                $this->reflect($class)
+                $this->methodReflector->reflect($class, '__construct')
             );
         }
 
         return $this->analysisStorage->get($class);
-    }
-
-    /**
-     * Retrieves a class analysis.
-     *
-     * @param string $class
-     *
-     * @return array
-     *
-     * @throws NonInstantiableClassException When the analysed class is not instantiable.
-     */
-    private function reflect(string $class): array
-    {
-        $reflection = new ReflectionClass($class);
-
-        if ($reflection->isInstantiable()) {
-            $constructor = $reflection->getConstructor();
-            $parameters = [];
-
-            if ($constructor !== null) {
-                foreach ($constructor->getParameters() as $parameter) {
-                    $type = $parameter->getType();
-                    $parameters[$parameter->getName()] = [
-                        'type' => $type !== null ?
-                            (array_key_exists($type->getName(), $this->translations)
-                                ? $this->translations[$type->getName()]
-                                : $type->getName())
-                            : 'mixed',
-                        'builtin' => $type !== null ? $type->isBuiltin(): true,
-                        'allowsNull' => $parameter->allowsNull(),
-                        'isOptional' => $parameter->isOptional(),
-                        'isVariadic' => $parameter->isVariadic(),
-                        'default' => $parameter->isDefaultValueAvailable()
-                            ? $parameter->getDefaultValue()
-                            : null,
-                    ];
-                }
-            }
-
-            return $parameters;
-        }
-
-        throw new NonInstantiableClassException(
-            $class
-        );
     }
 }