+++ src/Factory/ObjectFactory.php
@@ -1,13 +1,16 @@
 <?php
+
 /**
  * Copyright (C) GrizzIT, Inc. All rights reserved.
  * See LICENSE for license details.
  */
+
 namespace GrizzIt\ObjectFactory\Factory;
 
 use Throwable;
 use GrizzIt\ObjectFactory\Common\ClassAnalyserInterface;
 use GrizzIt\ObjectFactory\Common\ObjectFactoryInterface;
+use GrizzIt\ObjectFactory\Exception\MissingParameterException;
 use GrizzIt\ObjectFactory\Exception\CanNotCreateObjectException;
 use GrizzIt\ObjectFactory\Exception\InvalidParameterTypeException;
 
@@ -53,16 +56,40 @@ class ObjectFactory implements ObjectFactoryInterface
             );
         }
 
+        try {
+            $invokeParameters = $this->translate(
+                $config,
+                $parameters
+            );
+        } catch (Throwable $exception) {
+            throw new CanNotCreateObjectException(
+                $class,
+                $config,
+                $parameters,
+                $exception
+            );
+        }
+
+
+        return new $class(...$invokeParameters);
+    }
+
+    /**
+     * Converts the parameter order based on the configuration and returns a
+     * array which can be passed variadically.
+     *
+     * @return array
+     */
+    private function translate(array $config, array $parameters): array
+    {
         $invokeParameters = [];
         foreach ($config as $parameterName => $parameterConfig) {
             if (!isset($parameters[$parameterName])) {
-                if (!$parameterConfig['allowsNull']
-                && !$parameterConfig['isOptional']) {
-                    throw new CanNotCreateObjectException(
-                        $class,
-                        $parameterConfig,
-                        $parameters
-                    );
+                if (
+                    !$parameterConfig['allowsNull']
+                    && !$parameterConfig['isOptional']
+                ) {
+                    throw new MissingParameterException($parameterConfig);
                 }
 
                 $invokeParameters[] = $parameterConfig['default'];
@@ -71,43 +98,27 @@ class ObjectFactory implements ObjectFactoryInterface
             }
 
             if ($parameterConfig['isVariadic']) {
-                $variadicParameters = [];
                 foreach ($parameters[$parameterName] as $variadicParameter) {
-                    try {
-                        $invokeParameters[] = $this->parameterByType(
-                            $parameterConfig,
-                            $variadicParameter
-                        );
-                    } catch (Throwable $exception) {
-                        throw new CanNotCreateObjectException(
-                            $class,
-                            $parameterConfig,
-                            $parameters,
-                            $exception
-                        );
-                    }
+                    $invokeParameters[] = $this->parameterByType(
+                        $parameterConfig,
+                        $variadicParameter
+                    );
                 }
 
                 continue;
             }
 
-            try {
-                $invokeParameters[] = $this->parameterByType($parameterConfig, $parameters[$parameterName]);
-            } catch (Throwable $exception) {
-                throw new CanNotCreateObjectException(
-                    $class,
-                    $parameterConfig,
-                    $parameters,
-                    $exception
-                );
-            }
+            $invokeParameters[] = $this->parameterByType(
+                $parameterConfig,
+                $parameters[$parameterName]
+            );
         }
 
-        return new $class(...$invokeParameters);
+        return $invokeParameters;
     }
 
-    /**
-     * Validates and creates objects by it's type.
+        /**
+     * Checks the parameter type.
      *
      * @param array $parameterConfig
      * @param mixed $parameter
@@ -119,20 +130,27 @@ class ObjectFactory implements ObjectFactoryInterface
     private function parameterByType(array $parameterConfig, $parameter)
     {
         if ($parameterConfig['builtin']) {
-            if (gettype(
-                $parameter
-            ) === $parameterConfig['type']
-            || $parameterConfig['type'] === 'mixed') {
+            if (
+                gettype(
+                    $parameter
+                ) === $parameterConfig['type']
+                || $parameterConfig['type'] === 'mixed'
+            ) {
                 return $parameter;
             }
 
-            throw new InvalidParameterTypeException($parameterConfig, $parameter);
+            throw new InvalidParameterTypeException(
+                $parameterConfig,
+                $parameter
+            );
         }
 
-        if (!is_a(
-            $parameter,
-            $parameterConfig['type']
-        )) {
+        if (
+            !is_a(
+                $parameter,
+                $parameterConfig['type']
+            )
+        ) {
             if (is_array($parameter)) {
                 try {
                     return $this->create(
@@ -148,7 +166,10 @@ class ObjectFactory implements ObjectFactoryInterface
                 }
             }
 
-            throw new InvalidParameterTypeException($parameterConfig, $parameter);
+            throw new InvalidParameterTypeException(
+                $parameterConfig,
+                $parameter
+            );
         }
 
         return $parameter;