+++ CHANGELOG.md
@@ -4,8 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.1.0 - 2020-06-08
+### Added
+- Separate method reflector.
+
+### Refactored
+- Object factory to separate logic.
+
 ## 1.0.0 - 2020-04-09
 ### Added
 - Initial implementation of the object factory package.
 
-[1.0.0 > Unreleased](https://github.com/grizz-it/object-factory/compare/1.0.0...HEAD)
+- [1.1.0 > Unreleased](https://github.com/grizz-it/object-factory/compare/1.1.0...HEAD)
+- [1.0.0 > 1.1.0](https://github.com/grizz-it/object-factory/compare/1.0.0...1.1.0)